feat: add setting pages
